work from 12 27
All,
I am adding the PP with my analysis as well as the jupyter notebook with my update code. I have also included an updated resources file with some additional images that I generated. None were used in the PP but I think will be needed when we are graded.